+++ sample/src/main/java/me/drakeet/multitype/sample/MenuBaseActivity.java
@@ -4,7 +4,7 @@ import android.content.Intent;
 import android.support.v7.app.AppCompatActivity;
 import android.view.Menu;
 import android.view.MenuItem;
-import com.camnter.multitype.databinding.sample.normal.DataBindingNormalActivity;
+import com.camnter.multitype.databinding.sample.normal.BindingNormalActivity;
 import me.drakeet.multitype.sample.bilibili.BilibiliActivity;
 import me.drakeet.multitype.sample.communicate_with_binder.SimpleActivity;
 import me.drakeet.multitype.sample.multi_select.MultiSelectActivity;
@@ -51,8 +51,8 @@ public class MenuBaseActivity extends AppCompatActivity {
             case R.id.TestPayloadActivity:
                 intent.setClass(this, TestPayloadActivity.class);
                 break;
-            case R.id.DataBindingNormalActivity:
-                intent.setClass(this, DataBindingNormalActivity.class);
+            case R.id.BindingNormalActivity:
+                intent.setClass(this, BindingNormalActivity.class);
                 break;
             default:
                 return false;
