+++ sample/src/main/res/layout/activity_binding_normal_list.xml
@@ -52,7 +52,7 @@
             android:layout_centerInParent="true"
             android:gravity="center"
             android:orientation="horizontal"
-            android:visibility="@{viewModel.dataLoading?View.VISIBLE:View.GONE}">
+            android:visibility="@{collaborator.dataLoading?View.VISIBLE:View.GONE}">
 
             <ProgressBar
                 style="?android:attr/progressBarStyleSmall"
